Extract grammar graph from processor into grammar module
Move graph nodes, GrammarGraph, and related helpers out of processor.py into grammarinator.tool.grammar. Update processor.py and parser.py imports without changing behavior.
